requirements.txt
Flask
Pillow
rembg[cpu]